requirements.txt
networkx==2.5
tqdm
Pillow
numpy
opencv-python
scikit-image
scikit-learn
tensorflow_model_optimization
tensorboard